% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2alpha3.R
\name{convert2alpha3}
\alias{convert2alpha3}
\title{Convert input to alpha 3 code}
\usage{
convert2alpha3(country, nationality, alpha2, UN_code)
}
\arguments{
\item{country}{Vector of country to be converted}

\item{nationality}{Vector of nationalities to be converted}

\item{alpha2}{Vector of aplha 2 code to be converted}

\item{UN_code}{Vector of UN Code to be converted}

\item{alpha3}{Vector of alpha 3 code to be converted}
}
\value{
Vector of corresponding country names
}
\description{
This function takes a vector of nationalities and returns a vector of
corresponding country names.
}
\examples{
nationality <- c("American", "Canadian")
convert2alpha3(nationality)
}
\author{
Farhan Shah Jahan
}
