% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2nationality.R
\name{convert2nationality}
\alias{convert2nationality}
\title{Convert nationality to country name}
\usage{
convert2nationality(country, alpha2, alpha3, UN_code)
}
\arguments{
\item{country}{Vector of country to be converted}

\item{alpha2}{Vector of alpha 2 code to be converted}

\item{alpha3}{Vector of alpha 3 code to be converted}

\item{UN_code}{Vector of UN Code to be converted}

\item{nationality}{Vector of nationalities to be converted}
}
\value{
Vector of corresponding country names
}
\description{
This function takes a vector of nationalities and returns a vector of
corresponding country names.
}
\examples{
country <- c("Canada", "England")
convert2nationality(country)
}
\author{
Farhan Shah Jahan
}
