% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2un_code.R
\name{convert2un_code}
\alias{convert2un_code}
\title{Convert input to UN code}
\usage{
convert2un_code(country, alpha2, alpha3)
}
\arguments{
\item{country}{Vector of country to be converted}

\item{alpha2}{Vector of alpha 2 code to be converted}

\item{alpha3}{Vector of alpha 3 code to be converted}

\item{nationality}{Vector of nationalities to be converted}

\item{UN_code}{Vector of UN Code to be converted}
}
\value{
Vector of corresponding country names
}
\description{
This function takes a vector of nationalities and returns a vector of
corresponding country names.
}
\examples{
nationality <- c("American", "Canadian")
convert2un_code(nationality)
}
\author{
Farhan Shah Jahan
}
