% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nation_convert.R
\name{nation_convert}
\alias{nation_convert}
\title{Convert input to any target value inserted}
\usage{
nation_convert(origin, target, origin_data)
}
\arguments{
\item{origin}{category of original data}

\item{target}{category of target data}

\item{origin_data}{Vector of data that wanted to be transformed}
}
\value{
Vector of corresponding country names
}
\description{
This function takes input that can be Country, Nationality, Alpha2, Alpha3, UN_code and to be transformed to be any of them
}
\examples{
country_list <- c('India', 'Malaysia', 'Canada')
nation_convert(origin = 'Country', target = 'UN_code', origin_data = country_list)
}
\author{
Farhan Shah Jahan
}
